More work on C/PHP sep.

+++ projects/joern-fuzzyc/src/main/java/tools/argumentTainter/DefUseCFGPatcher.java
@@ -137,7 +137,7 @@ public class DefUseCFGPatcher
 				properties.put("code", link.symbol);
 
 				Node symbolNode = Neo4JDBInterface.addNode(properties);
-				toId = (Long) symbolNode.getId();
+				toId = symbolNode.getId();
 			}
 
 			RelationshipType relType = DynamicRelationshipType
